% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecmwfr2-package.R
\docType{package}
\name{ecmwfr2-package}
\alias{ecmwfr2}
\alias{ecmwfr2-package}
\title{ecmwfr2: Download ERA5 data with R.}
\description{
Download ERA5 data with R.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com} (\href{https://orcid.org/0000-0003-1836-8172}{ORCID})

}
\keyword{internal}
