% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilize.R
\name{getProcessInfo}
\alias{getProcessInfo}
\alias{write_url}
\title{get processing information from cds copernicus}
\usage{
getProcessInfo(outfile = "urls.txt", overwrite = TRUE, user = NULL)

write_url(d_url, outfile = "urls.txt", overwrite = TRUE)
}
\description{
get processing information from cds copernicus
}
