% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecmwfr_ERA5.R
\name{request_ERA5}
\alias{request_ERA5}
\title{request_ERA5}
\usage{
request_ERA5(
  varname = "temperature",
  years,
  transfer = TRUE,
  outdir = ".",
  ds = "reanalysis-era5-pressure-levels-monthly-means",
  product_type = "monthly_averaged_reanalysis"
)
}
\arguments{
\item{ds}{\itemize{
\item "reanalysis-era5-pressure-levels-monthly-means"
\item "reanalysis-era5-single-levels-monthly-means"
\item "reanalysis-era5-land-monthly-means"
}}
}
\description{
request_ERA5
}
